# docs/api/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "bibliothèque",
    "description" : "API de gestion de bibliothèque. Pour tester : bouton « Authorize », section « keycloak », saisir un compte (ex. admin / admin123 ou A1 / A1123), laisser client_id = bibliotheque-frontend, sans client_secret. Pour voir l'access token et le refresh token : POST /auth/token (à relancer quand l'access token expire)."
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "keycloak" : [ ]
  }, {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/loans" : {
      "get" : {
        "tags" : [ "Emprunts" ],
        "summary" : "Lister tous les emprunts (Admin / bibliothécaire)",
        "operationId" : "getAllBorrow",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BorrowResponse"
                  }
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Emprunts" ],
        "summary" : "Retourner un livre emprunté (un User uniquement le sien)",
        "operationId" : "returnBook",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ReturnBorrowRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BorrowResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Emprunts" ],
        "summary" : "Emprunter un livre (un User uniquement pour lui-même)",
        "operationId" : "borrowBook",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BorrowRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BorrowResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/books/{id}" : {
      "get" : {
        "tags" : [ "Livres" ],
        "summary" : "Obtenir un livre par son identifiant",
        "operationId" : "getBookById",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookResponse"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Livres" ],
        "summary" : "Modifier un livre existant",
        "operationId" : "updateBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Livres" ],
        "summary" : "Supprimer un livre",
        "operationId" : "deleteBook",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object",
                  "additionalProperties" : {
                    "type" : "boolean"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/users/{id}/anonymisation" : {
      "post" : {
        "tags" : [ "Données personnelles" ],
        "summary" : "Anonymiser un compte (droit à l'effacement, irréversible)",
        "operationId" : "anonymiser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/books" : {
      "get" : {
        "tags" : [ "Livres" ],
        "summary" : "Lister les livres (pagination)",
        "operationId" : "getAllBooks",
        "parameters" : [ {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "default" : "bookId"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageBookResponse"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Livres" ],
        "summary" : "Créer un nouveau livre",
        "operationId" : "createBook",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BookRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/BookResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reservations/{id}" : {
      "get" : {
        "tags" : [ "Réservations" ],
        "summary" : "Consulter une réservation (un ADHERENT uniquement la sienne, RS-03)",
        "operationId" : "consulter",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "404" : {
            "description" : "Réservation introuvable",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "200" : {
            "description" : "Réservation trouvée",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "Réservation d'un autre adhérent",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ReservationResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Réservations" ],
        "summary" : "Supprimer une réservation (bibliothécaire uniquement)",
        "operationId" : "supprimer",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Réservation supprimée"
          },
          "404" : {
            "description" : "Réservation introuvable"
          },
          "403" : {
            "description" : "Action réservée au bibliothécaire (RS-02)"
          }
        }
      }
    },
...
    "/api/v1/loans/user/{id}" : {
      "get" : {
        "tags" : [ "Emprunts" ],
        "summary" : "Lister les emprunts d'un utilisateur (un User uniquement les siens)",
        "operationId" : "booksBorrowedByUser",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BorrowResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/loans/book/{id}" : {
      "get" : {
        "tags" : [ "Emprunts" ],
        "summary" : "Historique des emprunts d'un livre (Admin / bibliothécaire)",
        "operationId" : "bookBorrowHistory",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/BorrowResponse"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "UserResponse" : {
        "type" : "object",
        "properties" : {
          "userId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "username" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "roles" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        }
      },
      "ReturnBorrowRequest" : {
        "required" : [ "borrowId" ],
        "type" : "object",
        "properties" : {
          "borrowId" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "BorrowResponse" : {
        "type" : "object",
        "properties" : {
          "borrowId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bookId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "issueDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "returnDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dueDate" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "BookRequest" : {
        "required" : [ "bookAuthor", "bookName", "noOfCopies" ],
        "type" : "object",
        "properties" : {
          "bookName" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "bookAuthor" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "bookGenre" : {
            "maxLength" : 255,
            "minLength" : 0,
            "type" : "string"
          },
          "noOfCopies" : {
            "minimum" : 0,
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "BookResponse" : {
        "type" : "object",
        "properties" : {
          "bookId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bookName" : {
            "type" : "string"
          },
          "bookAuthor" : {
            "type" : "string"
          },
          "bookGenre" : {
            "type" : "string"
          },
          "noOfCopies" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "ReservationResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int32"
          },
          "livreId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "livreNom" : {
            "type" : "string"
          },
          "adherentId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "adherentNom" : {
            "type" : "string"
          },
          "dateReservation" : {
            "type" : "string",
            "format" : "date-time"
          },
          "dateExpiration" : {
            "type" : "string",
            "format" : "date-time"
          },
          "statut" : {
            "type" : "string",
            "enum" : [ "EN_ATTENTE", "DISPONIBLE", "ANNULEE", "EXPIREE", "HONOREE" ]
          }
        }
      },
      "BorrowRequest" : {
        "required" : [ "bookId", "userId" ],
        "type" : "object",
        "properties" : {
          "bookId" : {
            "type" : "integer",
            "format" : "int32"
          },
          "userId" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "Pageable" : {
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "sort" : {
            "$ref" : "#/components/schemas/Sort"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paged" : {
            "type" : "boolean"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "Sort" : {
        "type" : "object",
        "properties" : {
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
...
      "PageBookResponse" : {
        "type" : "object",
        "properties" : {
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/BookResponse"
            }
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/Sort"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/Pageable"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    },
    "securitySchemes" : {
      "keycloak" : {
        "type" : "oauth2",
        "description" : "Connexion Keycloak (realm bibliotheque)",
        "flows" : {
          "password" : {
            "tokenUrl" : "http://localhost:9999/realms/bibliotheque/protocol/openid-connect/token",
            "scopes" : { }
          }
        }
      },
      "bearerAuth" : {
        "type" : "http",
        "description" : "Jeton d'accès Keycloak déjà obtenu (sans le préfixe Bearer)",
        "name" : "bearerAuth",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}